# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CEREBRO-RED v2 API",
    "description": "\n    # Autonomous Local LLM Red Teaming Suite - Research Edition\n    \n    ## Overview\n    CEREBRO-RED v2 implements the PAIR (Prompt Automatic Iterative Refinement) algorithm\n    for automated vulnerability discovery in Large Language Models.\n    \n    ## Features\n    - **Multi-Provider LLM Support**: Ollama, Azure OpenAI, OpenAI\n    - **8 Attack Strategies**: Obfuscation, Context Flooding, Roleplay Injection, etc.\n    - **LLM-as-a-Judge**: Semantic evaluation with 7 criteria\n    - **Real-Time Progress**: WebSocket streaming for live updates\n    - **Research-Grade Telemetry**: JSONL audit logs for analysis\n    \n    ## Authentication\n    All endpoints (except `/health`, `/docs`, `/metrics`) require API key authentication via `X-API-Key` header.\n    \n    ## Rate Limiting\n    Default: 60 requests per minute per IP address.\n    \n    ## References\n    - PAIR Paper: https://arxiv.org/abs/2310.08419\n    - GitHub: https://github.com/your-org/cerebro-red-v2\n    ",
    "version": "2.0.0",
    "contact": {
      "name": "CEREBRO-RED v2 Support",
      "url": "https://github.com/your-org/cerebro-red-v2/issues",
      "email": "support@cerebro-red.example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "paths": {
...
    "/api/v1/results/{experiment_id}": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Experiment Results",
        "description": "Get complete experiment results including all iterations and vulnerabilities.",
        "operationId": "get_experiment_results_api_v1_results__experiment_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/{experiment_id}/summary": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Experiment Summary",
        "description": "Get compact experiment result summary.",
        "operationId": "get_experiment_summary_api_v1_results__experiment_id__summary_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/{experiment_id}/export": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Export Experiment Results",
        "description": "Export experiment results in specified format (JSON, CSV, or PDF).",
        "operationId": "export_experiment_results_api_v1_results__experiment_id__export_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv|pdf)$",
              "default": "json",
              "title": "Format"
            }
          },
          {
            "name": "include_prompts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Prompts"
            }
          },
          {
            "name": "include_responses",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Responses"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/{experiment_id}/iterations/{iteration_id}": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Iteration Details",
        "description": "Get detailed information for a specific iteration.",
        "operationId": "get_iteration_details_api_v1_results__experiment_id__iterations__iteration_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          },
          {
            "name": "iteration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Iteration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:9000/api/v1",
...
    },
    {
...
    }
  ]
}